requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-bootstrap-v5==1.0.3
django-ckeditor==6.0.0
django-extensions==3.1.3
django-jet==1.0.8
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-taggit==1.4.0
djangorestframework==3.12.4
idna==2.10
mysqlclient==2.0.3
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
